requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
requests
twilio
google-cloud-storage
python-multipart
